requirements.txt
annotated-types==0.7.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
filelock==3.14.0
fiona==1.9.6
Flask==3.0.3
fonttools==4.51.0
fsspec==2024.5.0
gensim @ git+https://github.com/piskvorky/gensim.git@da2f38861ff097d2f9054db273b880b0ec98acdd
geopandas==0.14.4
huggingface-hub==0.23.2
idna==3.7
ijson==3.2.3
importlib_metadata==7.1.0
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mkl==2021.4.0
mpmath==1.3.0
mysql-connector-python==8.4.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
pyparsing==3.1.2
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.2
retrying==1.3.4
safetensors==0.4.3
scipy==1.13.1
setuptools==70.0.0
shapely==2.0.4
six==1.16.0
smart-open==7.0.4
spyder-kernels==2.5.1
stack-data==0.6.3
sympy==1.12
tbb==2021.12.0
tenacity==8.3.0
tokenizers==0.19.1
torch==2.3.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vaderSentiment==3.3.2
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.18.2
